Refactor test cases to reduce repetition
